Module draw::render

Renders Canvas objects to bytes, allows for saving to disk.

Modules

bitmap
Render Canvas to bitmap formats (PNG, JPG, etc...)

svg
Render Canvas to Scalable Vector Graphics (SVG) format

Traits

Renderer
A renderer takes a &Canvas and returns a Vec of bytes. This allows for drawings to be saved, passed to a conversion method, etc.

Functions

save
Saves a canvas to disk. Note: Will overwrite any existing file at the path you provide.