Function draw::render::save

pub fn save(canvas: &Canvas, path: &str, renderer: impl Renderer) -> Result<()>

Saves a canvas to disk. Note: Will overwrite any existing file at the path you provide.

canvas - Reference to the canvas you want to save

path - Full or relative path to where you want to save the file

renderer - An object that implements the Renderer trait