Trait draw::render::Renderer

pub trait Renderer {
    fn render(&self, canvas: &Canvas) -> Vec<u8>;
}

A renderer takes a &Canvas and returns a Vec of bytes. This allows for drawings to be saved, passed to a conversion method, etc.

Required methods

fn render(&self, canvas: &Canvas) -> Vec<u8>

Take a reference to a canvas, return a byte array

Loading content...